+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 1.0.3
+- FIXED: Not being able to dismount ladder in third person
+- FIXED: Hands not appearing when toggling back to first person
+- FIXED: Body not appearing on revive
+- FIXED: Being able to toggle third person in chat and terminal
+
+### 1.0.2
+- Added config value for StartEnabled (defaults to true)
+
 ### 1.0.1
 - FIXED: requiring crosshair.png
 	- Should work with Thunderstore now